requirements.txt
Flask==3.0.3
geopy==2.4.1
matplotlib==3.9.1
numpy==2.0.1
openpyxl==3.1.2
pandas==2.2.2
Pillow==10.4.0
protobuf>=3.19.5,<5.0.0dev 
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.31
tenacity==8.4.2
tqdm==4.66.1
gunicorn==21.2.0
google-generativeai==0.7.0
google-cloud-storage>=2.17.0
google-api-core>=2.11.0